refactor : add model directory
모델 디렉토리를 만들어 interface 를 임시 데이터와 분리함.
