+++ src/model/board.ts
@@ -1,0 +1,10 @@
+import { category } from './category';
+import { user } from './user';
+
+export interface board {
+  id: number;
+  title: string;
+  contents: string;
+  writerInfo: user;
+  category: category;
+}
