+++ src/model/category.ts
@@ -1,0 +1,5 @@
+export interface category {
+  id: number;
+  type: string;
+  priority?: number;
+}
