+++ src/model/user.ts
@@ -1,0 +1,5 @@
+export interface user {
+  id: string;
+  nickname: string;
+  userImageUrl?: string;
+}
